agent/pyproject.toml
[project]
name = "voiceorder-agent"
version = "0.1.0"
description = "VoiceOrder AI — Voice agent for restaurant phone ordering"
requires-python = ">=3.10"
dependencies = [
    "livekit-agents[codecs]~=1.4",
    "livekit-plugins-deepgram~=1.0",
    "livekit-plugins-openai~=1.0",
    "livekit-plugins-cartesia~=1.0",
    "livekit-plugins-silero~=1.0",
    "livekit-plugins-turn-detector~=1.0",
    "python-dotenv>=1.0",
    "livekit-plugins-anthropic~=1.4",
    "livekit-plugins-google~=1.4",
]

[project.scripts]
voiceorder = "voiceorder.agent:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/voiceorder"]
